+++ src/Controller/OperationalManagerController.php
@@ -19,68 +19,70 @@ class OperationalManagerController extends AbstractController
     /**
      * @Route("/", name="operational_manager_index", methods={"GET","POST"})
      */
-    public function index(OperationalManagerRepository $operationalManagerRepository,Request $request,PaginatorInterface $paginator): Response
+    public function index(OperationalManagerRepository $operationalManagerRepository, Request $request, PaginatorInterface $paginator): Response
     {
+        $this->denyAccessUnlessGranted('vw_opr_mng');
         $operationalManager = new OperationalManager();
         $form = $this->createForm(OperationalManagerType::class, $operationalManager);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $this->denyAccessUnlessGranted('ad_opr_mng');
             $entityManager = $this->getDoctrine()->getManager();
 
-             $isAlreadyAssigned=$operationalManagerRepository->findActive($form->getData()->getOperationalOffice(),$form->getData()->getManager());
-             $isActivePrincipal=$operationalManagerRepository->findActive($form->getData()->getOperationalOffice(),null);
-            
-             if ($isAlreadyAssigned) {
-                  $this->addFlash('danger'," operational manager  is already assigned to this office");
+            $isAlreadyAssigned = $operationalManagerRepository->findActive($form->getData()->getOperationalOffice(), $form->getData()->getManager());
+            $isActivePrincipal = $operationalManagerRepository->findActive($form->getData()->getOperationalOffice(), null);
 
-                     return $this->redirectToRoute('operational_manager_index');
-             }
-             if ($isActivePrincipal) {
-                  $this->addFlash('danger',"sorry unable to assign! the other operational manager  is already assigned to this office");
+            if ($isAlreadyAssigned) {
+                $this->addFlash('danger', " operational manager  is already assigned to this office");
 
-                     return $this->redirectToRoute('operational_manager_index');
-             }
+                return $this->redirectToRoute('operational_manager_index');
+            }
+            if ($isActivePrincipal) {
+                $this->addFlash('danger', "sorry unable to assign! the other operational manager  is already assigned to this office");
+
+                return $this->redirectToRoute('operational_manager_index');
+            }
 
             $entityManager->persist($operationalManager);
             $entityManager->flush();
 
-             $this->addFlash('success',"operational manager is assigned  successfuly");
+            $this->addFlash('success', "operational manager is assigned  successfuly");
 
             return $this->redirectToRoute('operational_manager_index');
         }
-         if($request->request->get('deactive')){
-            $operationalManager=$operationalManagerRepository->find($request->request->get('deactive'));
+        if ($request->request->get('deactive')) {
+            $this->denyAccessUnlessGranted('deact_opr_mng');
+            $operationalManager = $operationalManagerRepository->find($request->request->get('deactive'));
             $operationalManager->setIsActive(false);
-              $this->getDoctrine()->getManager()->flush();
-               $this->addFlash('success',"deactivated successfuly");
-              return $this->redirectToRoute('operational_manager_index');
-           
+            $this->getDoctrine()->getManager()->flush();
+            $this->addFlash('success', "deactivated successfuly");
+            return $this->redirectToRoute('operational_manager_index');
         }
-          if($request->request->get('active')){
-            $operationalManager=$operationalManagerRepository->find($request->request->get('active'));
-           
-              $isActivePrincipal=$operationalManagerRepository->findActive($operationalManager->getOperationalOffice(),null);
-               if ($isActivePrincipal) {
-                  $this->addFlash('danger',"sorry unable to activate! b/c this office has Active operational manager");
+        if ($request->request->get('active')) {
+            $this->denyAccessUnlessGranted('act_opr_mng');
+            $operationalManager = $operationalManagerRepository->find($request->request->get('active'));
 
-                     return $this->redirectToRoute('operational_manager_index');
-                  }
+            $isActivePrincipal = $operationalManagerRepository->findActive($operationalManager->getOperationalOffice(), null);
+            if ($isActivePrincipal) {
+                $this->addFlash('danger', "sorry unable to activate! b/c this office has Active operational manager");
+
+                return $this->redirectToRoute('operational_manager_index');
+            }
             $operationalManager->setIsActive(true);
-              $this->getDoctrine()->getManager()->flush();
-               $this->addFlash('success',"activated successfuly");
-              return $this->redirectToRoute('operational_manager_index');
-           
-         }
-           $data=$paginator->paginate(
-             $operationalManagerRepository->findAll(),
-             $request->query->getInt('page',1),
-             10
+            $this->getDoctrine()->getManager()->flush();
+            $this->addFlash('success', "activated successfuly");
+            return $this->redirectToRoute('operational_manager_index');
+        }
+        $data = $paginator->paginate(
+            $operationalManagerRepository->findAll(),
+            $request->query->getInt('page', 1),
+            10
 
         );
         return $this->render('operational_manager/index.html.twig', [
             'operational_managers' => $data,
-            'form'=>$form->createView()
+            'form' => $form->createView()
         ]);
     }
 
@@ -89,6 +91,7 @@ class OperationalManagerController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ad_opr_mng');
         $operationalManager = new OperationalManager();
         $form = $this->createForm(OperationalManagerType::class, $operationalManager);
         $form->handleRequest($request);
@@ -112,6 +115,7 @@ class OperationalManagerController extends AbstractController
      */
     public function show(OperationalManager $operationalManager): Response
     {
+        $this->denyAccessUnlessGranted('vw_opr_mng_dtl');
         return $this->render('operational_manager/show.html.twig', [
             'operational_manager' => $operationalManager,
         ]);
@@ -122,6 +126,7 @@ class OperationalManagerController extends AbstractController
      */
     public function edit(Request $request, OperationalManager $operationalManager): Response
     {
+        $this->denyAccessUnlessGranted('edt_opr_mng');
         $form = $this->createForm(OperationalManagerType::class, $operationalManager);
         $form->handleRequest($request);
 
@@ -142,7 +147,8 @@ class OperationalManagerController extends AbstractController
      */
     public function delete(Request $request, OperationalManager $operationalManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$operationalManager->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('dlt_opr_mng');
+        if ($this->isCsrfTokenValid('delete' . $operationalManager->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($operationalManager);
             $entityManager->flush();
